add google auth

+++ controllers/api/v1/posts_api.js
@@ -1,9 +1,10 @@
 const Post = require('../../../models/post');
 const Comment = require('../../../models/comment');
-
 module.exports.index = async function(req, res){
 
-    let posts =  await Post.find({})
+
+    let posts = await Post.find({})
+        .sort('-createdAt')
         .populate('user')
         .populate({
             path: 'comments',
@@ -12,9 +13,8 @@ module.exports.index = async function(req, res){
             }
         });
 
-
     return res.json(200, {
-        message: "Lists of posts",
+        message: "List of posts",
         posts: posts
     })
 }
@@ -25,24 +25,27 @@ module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
 
-       // if (post.user == req.user.id){
+        if (post.user == req.user.id){
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
 
 
+    
             return res.json(200, {
-                message: "Post and associated comments deleted" 
-            })
-        // }else{
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
+                message: "Post and associated comments deleted successfully!"
+            });
+        }else{
+            return res.json(401, {
+                message: "You cannot delete this post!"
+            });
+        }
 
     }catch(err){
-        console.log("##", err)
-        return res.json(500,{
-            messsage: "internal server error"
-        })
+        console.log('********', err);
+        return res.json(500, {
+            message: "Internal Server Error"
+        });
     }
+    
 }
